[lookup] move XOR table to tables mod
