+++ kimchi/src/circuits/polynomials/chacha.rs
@@ -146,48 +146,8 @@ use crate::circuits::{
     argument::{Argument, ArgumentType},
     expr::{constraints::boolean, prologue::*, ConstantExpr as C},
     gate::{CurrOrNext, GateType},
-    lookup::tables::LookupTable,
 };
 use ark_ff::{FftField, Field, Zero};
-
-/// The table ID associated with the XOR lookup table.
-pub const XOR_TABLE_ID: i32 = 0;
-
-/// The lookup table for 4-bit xor.
-/// Note that it is constructed so that (0, 0, 0) is the last position in the table.
-///
-/// This is because tables are extended to the full size of a column (essentially)
-/// by padding them with their final value. And, having the value (0, 0, 0) here means
-/// that when we commit to this table and use the dummy value in the `lookup_sorted`
-/// columns, those entries that have the dummy value of
-///
-/// 0 = 0 + joint_combiner * 0 + joint_combiner^2 * 0
-///
-/// will translate into a scalar multiplication by 0, which is free.
-pub fn xor_table<F: Field>() -> LookupTable<F> {
-    let mut data = vec![vec![]; 3];
-
-    // XOR for all possible four-bit arguments.
-    // I suppose this could be computed a bit faster using symmetry but it's quite
-    // small (16*16 = 256 entries) so let's just keep it simple.
-    for i in 0u32..=0b1111 {
-        for j in 0u32..=0b1111 {
-            data[0].push(F::from(i));
-            data[1].push(F::from(j));
-            data[2].push(F::from(i ^ j));
-        }
-    }
-
-    for r in &mut data {
-        r.reverse();
-        // Just to be safe.
-        assert!(r[r.len() - 1].is_zero());
-    }
-    LookupTable {
-        id: XOR_TABLE_ID,
-        data,
-    }
-}
 
 //
 // Implementation internals
